move query Location response DTO into common package
