+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/query/common/Location.kt
@@ -1,0 +1,6 @@
+package github.com.hukuta94.delivery.core.application.query.common
+
+data class Location(
+    val x: Int,
+    val y: Int,
+)
