+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/query/courier/response/Courier.kt
@@ -1,5 +1,6 @@
 package github.com.hukuta94.delivery.core.application.query.courier.response
 
+import github.com.hukuta94.delivery.core.application.query.common.Location
 import github.com.hukuta94.delivery.core.domain.courier.Transport
 import java.util.*
