--- core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/query/courier/response/Location.kt
@@ -1,6 +1,0 @@
-package github.com.hukuta94.delivery.core.application.query.courier.response
-
-data class Location(
-    val x: Int,
-    val y: Int,
-)
